requirements.txt
absl-py==0.10.0
aiohttp==3.7.3
antlr4-python3-runtime==4.8
anykeystore==0.2
astunparse==1.6.3
async-timeout==3.0.1
attributionpriors==0.1.3
attrs==20.2.0
blis==0.4.1
boto3==1.15.0
botocore==1.18.0
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptacular==1.5.5
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
defusedxml==0.7.1
editdistance==0.5.3
filelock==3.0.12
fsspec==0.8.5
future==0.18.2
gast==0.3.3
google-api-core==1.22.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.2
google-cloud-translate==3.0.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
hupper==1.10.2
hydra-core==1.0.1
idna==2.10
ignite==1.1.0
importlib-metadata==1.7.0
importlib-resources==3.0.0
iniconfig==1.0.1
jmespath==0.10.0
joblib==0.16.0
jsonnet==0.16.0
jsonpickle==1.4.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kmeans-pytorch==0.3
libcst==0.3.12
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
more-itertools==8.5.0
multidict==5.1.0
murmurhash==1.0.2
mypy-extensions==0.4.3
nltk==3.5
numpy==1.20.2
oauthlib==3.1.0
omegaconf==2.0.2
opt-einsum==3.3.0
optim==0.1.0
overrides==3.1.0
packaging==20.4
pandas==1.2.2
PasteDeploy==2.1.1
path-explain==0.0.5
pbkdf2==1.3
Pillow==7.2.0
plac==1.1.3
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.13.1
portalocker==2.0.0
preshed==3.0.2
proto-plus==1.10.0
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pyramid==2.0
pyramid-mailer==0.15.1
pytest==6.0.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytokenizations==0.8.1
pytorch-extension==0.1
pytorch-lightning==1.1.8
pytorch-pretrained-bert==0.6.2
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
repoze.sendmail==4.4.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacrebleu==1.4.14
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.4.1
sentencepiece==0.1.91
six==1.15.0
spacy==2.3.2
sparsemax==0.1.9
SQLAlchemy==1.4.3
srsly==1.0.2
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.8.0rc4
toml==0.10.1
torch==1.8.1
tqdm==4.49.0
transaction==3.0.1
transformers==3.0.0
translationstring==1.4
typing==3.6.4
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
velruse==1.1.1
venusian==3.0.0
wasabi==0.8.0
WebOb==1.8.7
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
yarl==1.6.3
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==5.3.0
zope.sqlalchemy==1.3
